rc: packages: added rich-minority package for modeline minification
rich-minority seems to be more permissive and versatile rather than
diminish in a couple of ways. Nevertheless, rich-minority lacks usable
lighters replacement functionality yet, so, until this functionality
implemented, it makes sense to use both packages, not at least because
it saves us from supporting a custom code for it.