rc: editing: navigate: fixed sudden readonly mode toggling on some buffers
it was introduced just a bit earlier with intent to make fiction books
buffers to be readonly to not suddenly edit book text. But the
implementation was too weak, so every buffer with "book" in name was
becoming readonly. As it seems safer to do such things by hands,
removing the respective code :)