customdefs: navigate: fixed #'open-urls-in-region
Seems there was a race condition when we called web browser to display
every next url from region, trying to search for url regexp and display
discovered url at once. It seems more correct to first collect *all*
urls from region and then open them one by one from a list. It still
presents a race condition as urls' order is not preserved, but we
eliminated the rare but very nasty case when the urls are found and open
and then the browser keeps opening new tabs, rendering itself (and
probably the whole system) unusable.