bootstrapping: do not explicitly load customization + remove on quit
Customize machinery tends to introduce hard-to-spot errors, yet all valuable
customizations could and should be persisted in main configuration. OTOH, rough
measures such as setting custom file to be null device, seem to be a bit inrusive
for the editing workflow. So, some kind of golden mean should be used.